fix: moved CityProjectData to DataProvider
* Moved `core.plotsystem.CityProjectData` to `core.data.CityProjectData`

* Make CityProjectData#fetchDataAsync no longer static

* Make CityProjectData constructor retrieve DataProvider for #fetchDataAync

* Removed constructor for making static CityProjectData cache,
  handled via the new constructor by setting expiry below zero instead.

* DataProviders is now initialized with cacheDurationMinute

* DataProvider#getCityProjectData: get city project data cache
