+++ src/main/java/com/alpsbte/plotsystemterra/core/api/DataProviderAPI.java
@@ -24,6 +24,7 @@
 
 package com.alpsbte.plotsystemterra.core.api;
 
+import com.alpsbte.plotsystemterra.core.data.CityProjectData;
 import com.alpsbte.plotsystemterra.core.data.CityProjectDataProvider;
 import com.alpsbte.plotsystemterra.core.data.DataProvider;
 import com.alpsbte.plotsystemterra.core.data.PlotDataProvider;
@@ -31,13 +32,15 @@ import com.alpsbte.plotsystemterra.core.data.PlotDataProvider;
 public class DataProviderAPI implements DataProvider {
     private final CityProjectDataProvider cityProjectDataProvider;
     private final PlotDataProvider plotDataProvider;
+    private final CityProjectData cityProjectData;
 
-    public DataProviderAPI() {
+    public DataProviderAPI(int expiryMinute) {
         // Initialize API constants
         ApiConstants.updateApiConstants();
 
         cityProjectDataProvider = new CityProjectDataProviderAPI();
         plotDataProvider = new PlotDataProviderAPI();
+        cityProjectData = new CityProjectData(cityProjectDataProvider, expiryMinute);
     }
 
     @Override
@@ -48,6 +51,11 @@ public class DataProviderAPI implements DataProvider {
     @Override
     public PlotDataProvider getPlotDataProvider() {
         return plotDataProvider;
+    }
+
+    @Override
+    public CityProjectData getCityProjectData() {
+        return cityProjectData;
     }
 
     @Override
