+++ src/main/java/com/alpsbte/plotsystemterra/core/data/DataProvider.java
@@ -27,6 +27,7 @@ package com.alpsbte.plotsystemterra.core.data;
 public interface DataProvider {
     CityProjectDataProvider getCityProjectDataProvider();
     PlotDataProvider getPlotDataProvider();
+    CityProjectData getCityProjectData();
 
     void shutdown();
 }
