+++ src/main/java/com/alpsbte/plotsystemterra/core/database/DataProviderSQL.java
@@ -27,6 +27,7 @@ package com.alpsbte.plotsystemterra.core.database;
 import com.alpsbte.alpslib.io.database.DatabaseConfigPaths;
 import com.alpsbte.alpslib.io.database.DatabaseConnection;
 import com.alpsbte.plotsystemterra.PlotSystemTerra;
+import com.alpsbte.plotsystemterra.core.data.CityProjectData;
 import com.alpsbte.plotsystemterra.core.data.CityProjectDataProvider;
 import com.alpsbte.plotsystemterra.core.data.DataProvider;
 import com.alpsbte.plotsystemterra.core.data.PlotDataProvider;
@@ -38,14 +39,16 @@ import static net.kyori.adventure.text.Component.text;
 public class DataProviderSQL implements DataProvider {
     private final CityProjectDataProvider cityProjectDataProvider;
     private final PlotDataProvider plotDataProvider;
+    private final CityProjectData cityProjectData;
 
-    public DataProviderSQL(@NotNull PlotSystemTerra plugin, @NotNull Component successPrefix) throws ClassNotFoundException {
+    public DataProviderSQL(@NotNull PlotSystemTerra plugin, @NotNull Component successPrefix, int expiryMinute) throws ClassNotFoundException {
         // Initialize database connection
         DatabaseConnection.initializeDatabase(DatabaseConfigPaths.getConfig(plugin.getConfig()), true);
         plugin.getComponentLogger().info(successPrefix.append(text("Successfully initialized database connection.")));
 
         cityProjectDataProvider = new CityProjectDataProviderSQL();
         plotDataProvider = new PlotDataProviderSQL();
+        cityProjectData = new CityProjectData(cityProjectDataProvider, expiryMinute);
     }
 
     @Override
@@ -62,5 +65,10 @@ public class DataProviderSQL implements DataProvider {
     @Override
     public PlotDataProvider getPlotDataProvider() {
         return plotDataProvider;
+    }
+
+    @Override
+    public CityProjectData getCityProjectData() {
+        return cityProjectData;
     }
 }
